rearranging methods and classes
